% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{frequencies}
\alias{frequencies}
\title{Acquire frequency values for any quantile of energy}
\usage{
frequencies(spec, EQ = c(0.05, 0.15, 0.3, 0.5, 0.7, 0.85, 0.95),
  f = NULL, mel = FALSE, flim = NULL)
}
\arguments{
\item{spec}{a data set resulting of a spectral analysis obtained with functions \code{\link{spec}} or \code{\link{meanspec}} (\code{\link{seewave}} package) (not in dB)}

\item{EQ}{a vector of energy quantiles to be calculated (with 0 < \code{EQ} < 1). By default: \code{EQ = c(0.05, 0.15, 0.3, 0.5, 0.7, 0.85, 0.95)}}

\item{f}{sampling frequency of \code{spec} (in Hz)}

\item{mel}{a logical, if \code{TRUE} the (htk-)mel scale is used.}

\item{flim}{a vector of length 2 to specify the frequency limits of the analysis (in kHz)}
}
\description{
Calculation of frequency values that represent as many energy quantiles as desired. Codes adapted from \code{\link{specprop}} (\code{\link{seewave}} package).
}
\note{
Codes of \code{frequencies} were adapted from \code{\link{specprop}} (\code{\link{seewave}} package) so that it returns a \code{data.frame} with frequency values that represent as many energy quantiles (\code{EQ}) as desired.

In addition, arguments \code{spec}, \code{f}, \code{mel} and \code{flim} are the same as in \code{\link{specprop}}.
}
\examples{
library(seewave)

data(cuvieri)
sample.sound <- spec(cuvieri)

# assign result to object
sample.freq <- frequencies(sample.sound)

# print result
print(sample.freq)

# a given energy quantile
sample.freq2 <- frequencies(sample.sound, EQ = 0.2)

# multiple energy quantiles
sample.freq3 <- frequencies(sample.sound, EQ = c(0.2, 0.4, 0.6, 0.8))

# multiple energy quantiles within a predefined frequency bandwidth
sample.freq4 <- frequencies(sample.sound, EQ = 0.5, f=44100, flim = c(2.5, 10))


}
\references{
Rocha, P. & Romano, P. (\emph{in prep}) The shape of sound: A new \code{R} package that crosses the bridge between Geometric Morphometrics and Bioacoustics.

Sueur J., Aubin T., Simonis C. (2008). Seewave: a free modular tool for sound analysis and synthesis. \emph{Bioacoustics, 18}: 213-226
}
\seealso{
\code{\link{spectro}}, \code{\link{specprop}}, \code{\link{spec}}, \code{\link{meanspec}}, \code{\link{seewave}}

Useful links:
\itemize{
  \item{\url{https://github.com/p-rocha/SoundShape}}
  \item{Report bugs at \url{https://github.com/p-rocha/SoundShape/issues}}}
}
