% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raven.to.wave.R
\name{raven.to.wave}
\alias{raven.to.wave}
\title{Export sample \code{".wav"} files using selections from Raven Pro software.}
\usage{
raven.to.wave(
  orig.wav = NULL,
  raven.at = orig.wav,
  wav.samples = "wav samples",
  max.dur = NULL
)
}
\arguments{
\item{orig.wav}{filepath to the folder where original \code{".wav"} files are stored. Should be presented between quotation marks. By default: \code{orig.wav = NULL} (i.e. user must specify the filepath to \code{".wav"} files)}

\item{raven.at}{filepath to the folder where selection tables from \href{https://ravensoundsoftware.com/software/raven-pro/}{Raven Pro} software are stored. Should be presented between quotation marks. File name should end with \code{"selections.txt"}. By default: \code{raven.at = orig.wav} (i.e. raven tables stored in the same folder as original  \code{".wav"} files)}

\item{wav.samples}{name of the folder where new \code{".wav"} files will be stored. Should be presented between quotation marks. By default: \code{wav.samples = "wav samples"}}

\item{max.dur}{optional argument to specify intended duration of sample \code{".wav"} files. If \code{max.dur=NULL}, function \code{raven.to.wav} will define the duration as 20\% longer than the maximum duration of selections from Raven Pro software. By default: \code{max.dur=NULL}}
}
\description{
create one \code{".wav"} file for each selection created using \href{https://ravensoundsoftware.com/software/raven-pro/}{Raven Pro} software, which is commonplace in bioacoustical analysis. Each selection (i.e. line in table) should represent a each acoustic unit within the sample study.
}
\examples{
\donttest{

# Write example

}

}
\references{
MacLeod, N., Krieger, J. & Jones, K. E. (2013). Geometric morphometric approaches to acoustic signal analysis in mammalian biology. \emph{Hystrix, the Italian Journal of Mammalogy, 24}(1), 110-125.

Rocha, P. & Romano, P. (2021) The shape of sound: A new \code{R} package that crosses the bridge between Bioacoustics and Geometric Morphometrics. \emph{Methods in Ecology and Evolution, 12}(6), 1115-1121.
}
\seealso{
\code{\link{align.wave}}

Useful links:
\itemize{
\item{\url{https://github.com/p-rocha/SoundShape}}
\item{Report bugs at \url{https://github.com/p-rocha/SoundShape/issues}}
\item{Raven Pro software \url{https://ravensoundsoftware.com/software/raven-pro/}}}
}
\author{
Pedro Rocha
}
