\docType{package}
\name{SoundShape}
\alias{SoundShape}
\alias{SoundShape-package}

\title{Sound Waves Onto Morphometric Data}

\description{A set of functions that crosses the bridge between Bioacoustics and Geometric Morphometrics.}

\itemize{
  \item{\code{\link{eigensound}} }
  \item{\code{\link{align.wave}} }
  \item{\code{\link{pca.plot}} }
}

\seealso{
\code{\link{geomorph}}, \code{\link{seewave}}

Useful links:
\itemize{
  \item{\url{https://github.com/p-rocha/SoundShape}}
  \item{Report bugs at \url{https://github.com/p-rocha/SoundShape/issues}}}
}

\author{
Pedro Rocha \email{p.rocha1990@gmail.com}
}

\references{
Rocha, P. & Romano, P. (\emph{in prep}) The shape of sound: A new \code{R} package that crosses the bridge between Geometric Morphometrics and Bioacoustics.
}
