% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kroyeri.R
\docType{data}
\name{kroyeri}
\alias{kroyeri}
\title{Advertisement call of \emph{Physalaemus kroyeri}}
\format{An object of the class \code{"Wave"} (\code{\link{tuneR}} package.}
\source{
Original recording housed at Fonoteca Neotropical Jacques Vielliard (FNJV-0032047). Recorded by Werner Bokermann.
}
\usage{
data(kroyeri)
}
\description{
Recording of \emph{Physalaemus kroyeri} (Amphibia, Anura, Leptodactylidae) advertisement call containing three notes emitted in a sequence. Edited from original \code{Wave} file for optimal sinal to noise ratio and reduced time duration.
}
\details{
Duration = 3.91 s. Sampling Frequency = 44100 Hz. Air temperature 24ºC.

Recorded at Ilhéus Municipality, Bahia State, Brazil, on 05 August 1972.
}
\examples{
data(kroyeri)

seewave::oscillo(kroyeri, f=22050)
seewave::spectro(kroyeri, f=22050)
}
\keyword{datasets}
